removing unnided vars and moving unsplash wallpaper module
